mission: only drop mission messages for now
To make test results more reproducible it makes sense to only filter out
mission specific messages. That way we don't depend on the timing and
rate of the rest of the messages and should get consistent drops for
messages which matter for the mission transfer.

Note that the MISSION_ACK message is not dropped for now because there
is no retransmission in the protocol for it.