main: fix unclean destruction
By making the promise a std::shared_ptr we don't run into the case where
it already has gone out of scope by the time the callback is called.

Also, we can unregister the component callback, so we don't get notified
anymore. However, we can't be sure that the component discovered
callback isn't called twice in rapid succession, so we still need to
catch the exception.